+++ source/pipes/formatting/date-time-to-relative-calendar.ts
@@ -1,0 +1,13 @@
+import { Pipe, PipeTransform } from '@angular/core'
+import { DateTime, ToRelativeUnit } from 'luxon'
+
+@Pipe({
+  name: 'dateTimeToRelativeCalendar'
+})
+export class DateTimeToRelativeCalendarPipe implements PipeTransform {
+  transform (value: DateTime, unit?: ToRelativeUnit): string {
+    return value.toRelativeCalendar({
+      unit
+    }) as string
+  }
+}