+++ source/pipes/formatting/date-time-to-relative.ts
@@ -1,0 +1,14 @@
+import { Pipe, PipeTransform } from '@angular/core'
+import { DateTime, ToRelativeUnit } from 'luxon'
+
+@Pipe({
+  name: 'dateTimeToRelative'
+})
+export class DateTimeToRelativePipe implements PipeTransform {
+  transform (value: DateTime, unit?: ToRelativeUnit, style?: 'long' | 'short' | 'narrow'): string {
+    return value.toRelative({
+      style,
+      unit
+    }) as string
+  }
+}