+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project are documented here. The format is based on 
 
 ## Unreleased
 
+-   Added DateTime -> relative pipes.
 -   Configured build to include ES5 and ES2015 outputs.
 
 ## 1.0.0 (2018-11-28)